+++ crates/app/src/channel/mod.rs
@@ -56,14 +56,12 @@ pub mod traits;
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-mod turn_feedback;
 #[cfg(feature = "channel-twitch")]
 mod twitch;
 #[cfg(feature = "channel-twitch")]
 mod twitch_command;
 #[cfg(feature = "channel-webhook")]
 mod webhook;
-mod webhook_auth;
 #[cfg(feature = "channel-wecom")]
 mod wecom;
 #[cfg(feature = "channel-whatsapp")]
@@ -119,7 +117,7 @@ pub use tlon_command::run_tlon_send;
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-pub use turn_feedback::ChannelTurnFeedbackPolicy;
+pub use runtime::turn_feedback::ChannelTurnFeedbackPolicy;
 
 mod types;
 pub use types::ChannelOutboundTargetKind as ChannelCatalogTargetKind;
