+++ crates/app/src/channel/runtime/mod.rs
@@ -1,3 +1,4 @@
 // crates/app/src/channel/runtime/mod.rs
 
 pub(in crate::channel) mod http;
+pub(in crate::channel) mod turn_feedback;
